Updated @name documentation. Closes #25.
The new documentation makes it clear @name must be specified in a separate comment, not mixed in with method or property comment.

It might be nicer to allow users freedom of choosing where to put section titles, but it would make the code more complicated, so for the moment the simpler solution was chosen.